feat(abstract-utxo): add signTransaction tests
Add comprehensive tests for signTransaction functionality, covering:
- PSBT with taprootKeyPathSpend inputs
- PSBT without taprootKeyPathSpend inputs
- Network transactions
- Error handling for cache misses and unsupported locking scripts
- Validation of signature and nonce counts

Issue: BTC-2866
